+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNEXT
 
+- feat: custom `depthLimit` (circular references) [PR #29](https://github.com/Urigo/SOFA/pull/29)
+
 ### v0.2.3
 
 - fix: prevent circular references [PR #23](https://github.com/Urigo/SOFA/pull/23)